Support ZIP64

+++ src/StreamingZipReader/StreamingZipEntry.cs
@@ -1,3 +1,3 @@
 namespace Wololo.StreamingZipReader;
 
-public readonly record struct StreamingZipEntry(string Name, uint Crc32, uint CompressedLength, uint Length);
+public readonly record struct StreamingZipEntry(string Name, uint Crc32, long CompressedLength, long Length);
